+++ config/config.php
@@ -6,6 +6,9 @@
  * @var array<string, mixed> $config
  * @var \AMPBoard\Database\ConnectionFactory $database
  * @var \AMPBoard\Ui\Renderer $ui
+ * @var \AMPBoard\Apache\CommandRunner $apacheCommands
+ * @var \AMPBoard\Apache\Controller $apacheControl
+ * @var \AMPBoard\Apache\VhostCatalog $vhosts
  */
 
 require_once __DIR__ . '/autoload.php';
@@ -19,3 +22,10 @@ $profiles = new \AMPBoard\Config\ProfileRepository( __DIR__, $cipher, null, \AMP
 $config = ( new \AMPBoard\Config\Loader( __DIR__, $profiles ) )->load( true );
 $database = new \AMPBoard\Database\ConnectionFactory( $config['db'] );
 $ui = new \AMPBoard\Ui\Renderer( $config, $database );
+
+$apacheCommands = new \AMPBoard\Apache\ShellCommandRunner();
+$apacheControl = new \AMPBoard\Apache\Controller( $config['paths']['apache'], PHP_OS_FAMILY, $apacheCommands );
+$vhosts = new \AMPBoard\Apache\VhostCatalog( $config['paths']['apache'], [
+	getenv( 'WINDIR' ) ? getenv( 'WINDIR' ) . '/System32/drivers/etc/hosts' : '',
+	'/etc/hosts',
+] );